Major refactor
After some discussion and some pep8 related warnings, i went for a major refactor of the script in terms of it's structure to make it less complex, and hopefully further improve readability.
